requirements.txt
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.1.0
cachetools==4.2.2
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.0
crypto==1.4.1
cryptography==42.0.2
cycler==0.11.0
et-xmlfile==1.1.0
Fiona==1.9.4.post1
fonttools==4.41.1
fuzzywuzzy==0.18.0
geopandas==0.13.2
idna==3.6
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
jieba==0.42.1
-e git+https://github.com/kpu/kenlm.git@5bf7b46558e1c5595bf3b8c9b0b1f9d8d257040a#egg=kenlm
kiwisolver==1.4.4
Levenshtein==0.21.1
loguru==0.7.0
matplotlib==3.7.2
MouseInfo==0.1.3
mpmath==1.3.0
Naked==0.1.32
numpy==1.24.2
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pinyin==0.4.0
Pinyin2Hanzi==0.1.1
protobuf==4.25.3
psutil==5.9.4
PyAutoGUI==0.9.54
pycorrector==0.5.0
pycparser==2.21
pycryptodome==3.20.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==24.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pypinyin==0.49.0
pyproj==3.6.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-Levenshtein==0.21.1
python-telegram-bot==13.7
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
rapidfuzz==3.1.2
redis==5.0.1
requests==2.31.0
shapely==2.0.1
shellescape==3.8.1
six==1.16.0
tornado==6.4
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
zipp==3.16.2
spacy==3.7.4
firebase-admin==6.5.0